\name{hydroPSO-internal}
\title{Internal hydroPSO objects}
\alias{ParameterValues2InputFiles}
\alias{read.ParameterRanges}
\alias{rLHS}
\alias{hydromod.eval.SA}
\alias{hydromodInR.eval}
\description{Internal hydroPSO objects.}
\details{These are not to be called by the user.}
\keyword{internal}
\value{It does not return a value. It is only sued to declare internal package objects, which are not to be called by the user} 
