%% Part of the hydroPSO R package, http://www.rforge.net/hydroPSO/ ; 
%%                                 http://cran.r-project.org/web/packages/hydroPSO
%% Copyright 2011-2026 Mauricio Zambrano-Bigiarini & Rodrigo Rojas
%% Distributed under GPL 2 or later

\name{read_best}
\Rdversion{1.1}
\alias{read_best}

\title{
Reading the 'BestParameterSet.txt' output file
}
\description{
This function reads the contents of the the \sQuote{BestParameterSet.txt} output file, which stores the best parameter set and its corresponding goodness-of-fit value found during the optimisation
}
\usage{
read_best(file="BestParameterSet.txt", verbose=TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
character, name (including path) of the output file with the best parameter set  and its corresponding best fitness value found during the optimisation
}
  \item{verbose}{
logical, if TRUE, progress messages are printed
}
}

\value{
a list with the following three elements:

\item{best.part.number}{numeric of length 1, with the number of PSO particle that achieved the best optimisation value.} \cr 

\item{best.param.values}{numeric with the parameter values of PSO particle that achieved the best optimisation value.} \cr 

\item{best.param.gof}{numeric of length one with the GoF of the PSO particle that achieved the best optimisation value.}
}

\seealso{
\code{\link{read_results}}, \code{\link{plot_results}}
}
\examples{
\donttest{
local({

# Setting the user temporal directory as working directory
oldwd <- getwd()      
on.exit(setwd(oldwd), add = TRUE) 
setwd(tempdir())

# Number of dimensions to be optimised
D <- 4

# Boundaries of the search space (Sphere test function)
lower <- rep(-100, D)
upper <- rep(100, D)

# Setting the seed
set.seed(100)

# Runing PSO with the 'Sphere' test function, writting the results to text files
pso.dir <- file.path(tempdir(), "PSO.out")
hydroPSO(fn=sphere, lower=lower, upper=upper,
        control=list(maxit=100, write2disk=TRUE, drty.out=pso.dir,
                     plot=TRUE)  ) 
  
# Reading the best parameter set and its corresponding gof found by hydroPSO
read_best(file=file.path(pso.dir, "BestParameterSet.txt"))

}) # local END
} # donttest END
}
%\keyword{graph}
\keyword{manip}
