%% File verification.Rd
%% Part of the hydroPSO R package, https://github.com/hzambran/hydroPSO
%%                                 http://cran.r-project.org/web/packages/hydroPSO
%%                                 http://www.rforge.net/hydroPSO/
%% Copyright 2011-2026 Mauricio Zambrano-Bigiarini & Rodrigo Rojas
%% Distributed under GPL 2 or later

\name{verification}
\alias{verification}

\title{
verification
}
\description{
Run the model and get a goodness-of-fit value by comparing the simulated values against observations for the optimum parameter set found by optimisation
}
\usage{
verification(fn="hydromod", par, ..., control=list(), 
             model.FUN=NULL, model.FUN.args=list() )
}
\arguments{
  \item{fn}{
character, name of a valid R function to be optimised or character value \kbd{'hydromod'}. When \code{fn='hydromod'} the algorithm uses \code{model.FUN} and \code{model.FUN.args} to extract the values simulated by the model and to compute its corresponding goodness-of-fit function. When \code{fn!='hydromod'} the algorithm uses the value(s) returned by \code{fn} as both model output and its corresponding goodness-of-fit. \cr

When \code{fn='hydromod'} the algorithm will optimise the model defined by \code{model.FUN} and \code{model.args}
}
\item{\dots}{
OPTIONAL. Only used when \code{fn!='hydromod' & fn!='hydromodInR'}. \cr

further arguments to be passed to \code{fn}.
}
  \item{par}{
numeric, or matrix/data.frame with the parameter sets that will be used for verification. \cr

Parameter sets in \code{par} must be stored by row, i.e., each different row represents a different parameter set
}
  \item{control}{
a list of control parameters. See \sQuote{Details}.
}
  \item{model.FUN}{
OPTIONAL. Only used when \code{fn='hydromod'}. \cr

character, valid R function representing the model code to be calibrated/optimised
} 
  \item{model.FUN.args}{
OPTIONAL. Only used when \code{fn='hydromod'}. \cr

list with the arguments to be passed to \code{model.FUN}
} 

}
\details{
The \code{control} argument is a list that can supply any of the following components:
  \describe{  
      
  \item{drty.in}{
character, path to the directory storing the input files required for PSO, i.e. \sQuote{ParamRanges.txt} and \sQuote{ParamFiles.txt}.
}
  \item{drty.out}{
character, path to the directory storing the output files generated by \kbd{hydroPSO}.
}
  \item{digits}{
OPTIONAL. Only used when \code{write2disk=TRUE}. \cr

numeric, number of significant digits used for writing the outputs in scientific notation
}
  \item{gof.name}{
character, ONLY used for identifying the goodness-of-fit of each model run and writing it to the \kbd{Verification-logfile.txt}, \kbd{Verification-ParamValues.txt} output files.
}
  \item{MinMax}{
character, indicates whether the optimum value for the analysed problem corresponds to the minimum or maximum of the the objective function. It is used to select the \sQuote{best} parameter set. \cr 

Valid values are in: \code{c('min', 'max')}
}
  \item{do.plots}{
logical, if TRUE a PNG plot with the comparison between observed and simulated values should be produced.
}
  \item{write2disk}{
logical, indicates if the output files will be written to the disk.
}
  \item{verbose}{
logical, if TRUE progress messages are printed.
}
  }
}

\value{
A list of four elements: \cr 

\item{gofs}{numeric with the goodness-of-fit values corresponding to each one of the parameter sets provided in \code{par}}.

\item{model.values}{data.frame with the model outputs corresponding to each one of the parameter sets provided in \code{par}}.

\item{best.param}{numeric with the parameter values of the "best" paraemter set found during the verification period}.

\item{best.gof}{numeric with the goodness-of-fit of the "best" parameter set found during the verification period}

}

\author{
Mauricio Zambrano-Bigiarini, \email{mzb.devel@gmail.com}
}

\seealso{
\code{\link{hydromod}}, \code{\link{hydromodInR}}
}

\keyword{optimize}
\keyword{files}% __ONLY ONE__ keyword per line
