%% File verification.Rd
%% Part of the hydroPSO R package, https://github.com/hzambran/hydroPSO
%%                                 http://cran.r-project.org/web/packages/hydroPSO
%%                                 http://www.rforge.net/hydroPSO/
%% Copyright 2011-2026 Mauricio Zambrano-Bigiarini & Rodrigo Rojas
%% Distributed under GPL 2 or later

\name{verification}
\alias{verification}

\title{
verification
}
\description{
Run the model and get a goodness-of-fit value by comparing the simulated values against observations for the optimum parameter set found by optimisation
}
\usage{
verification(fn="hydromod", par, ..., control=list(), 
             model.FUN=NULL, model.FUN.args=list(),
             change.type="repl", refValue=NULL )
}
\arguments{
  \item{fn}{
character, name of a valid R function to be optimised or character value \kbd{'hydromod'}. When \code{fn='hydromod'} the algorithm uses \code{model.FUN} and \code{model.FUN.args} to extract the values simulated by the model and to compute its corresponding goodness-of-fit function. When \code{fn!='hydromod'} the algorithm uses the value(s) returned by \code{fn} as both model output and its corresponding goodness-of-fit. \cr

When \code{fn='hydromod'} the algorithm will optimise the model defined by \code{model.FUN} and \code{model.args}
}
\item{\dots}{
OPTIONAL. Only used when \code{fn!='hydromod' & fn!='hydromodInR'}. \cr

further arguments to be passed to \code{fn}.
}
  \item{par}{
numeric, or matrix/data.frame with the parameter sets that will be used for verification. \cr

Parameter sets in \code{par} must be stored by row, i.e., each different row represents a different parameter set
}
  \item{control}{
a list of control parameters. See \sQuote{Details}.
}
  \item{model.FUN}{
OPTIONAL. Only used when \code{fn='hydromod'}. \cr

character, valid R function representing the model code to be calibrated/optimised
} 
  \item{model.FUN.args}{
OPTIONAL. Only used when \code{fn='hydromod'}. \cr

list with the arguments to be passed to \code{model.FUN}
} 
  \item{change.type}{
OPTIONAL. Used only when \code{fn='hydromodInR'}. Character of length 1 or length equal to the number of parameters, indicating how the parameter sets in \code{par} are converted to the parameter values passed to \code{model.FUN}. Valid values are \code{"repl"} for replacement, \code{"addi"} for additive changes, and \code{"mult"} for multiplicative changes. A single value is recycled to all parameters. The default \code{"repl"} preserves the previous behaviour.
}
  \item{refValue}{
OPTIONAL. Used only when \code{fn='hydromodInR'} and any element of \code{change.type} is \code{"addi"} or \code{"mult"}. Numeric reference value of length 1 or length equal to the number of parameters, used to compute the parameter values passed to \code{model.FUN}. Only parameters using \code{"addi"} or \code{"mult"} require and use their corresponding reference value.
}

}
\details{
The \code{control} argument is a list that can supply any of the following components:
  \describe{  
      
  \item{drty.in}{
character, path to the directory storing the input files required for PSO, i.e. \sQuote{ParamRanges.txt} and \sQuote{ParamFiles.txt}.
}
  \item{drty.out}{
character, path to the directory storing the output files generated by \kbd{hydroPSO}.
}
  \item{digits}{
OPTIONAL. Only used when \code{write2disk=TRUE}. \cr

numeric, number of significant digits used for writing the outputs in scientific notation
}
  \item{gof.name}{
character, ONLY used for identifying the goodness-of-fit of each model run and writing it to the \kbd{Verification-logfile.txt}, \kbd{Verification-ParamValues.txt} output files.
}
  \item{MinMax}{
character, indicates whether the optimum value for the analysed problem corresponds to the minimum or maximum of the the objective function. It is used to select the \sQuote{best} parameter set. \cr 

Valid values are in: \code{c('min', 'max')}
}
  \item{do.plots}{
logical, if TRUE a PNG plot with the comparison between observed and simulated values should be produced.
}
  \item{write2disk}{
logical, indicates if the output files will be written to the disk.
}
  \item{verbose}{
logical, if TRUE progress messages are printed.
}
  \item{parallel}{
character, indicates how to parallelise \sQuote{verification} (to be precise, only the evaluation of the objective function \code{fn} is parallelised). Valid values are: \cr

-)\kbd{none}: no parallelisation is made (this is the default value) \cr

-)\kbd{multicore}: DEPRECATED. The former \pkg{multicore} package is no longer available from CRAN. For backward compatibility, \code{parallel="multicore"} is automatically changed to \code{parallel="parallel"} on Unix-like systems and to \code{parallel="parallelWin"} on Microsoft Windows. \cr

-)\kbd{parallel}: parallel computations using a \sQuote{FORK} cluster created with \code{\link[parallel]{makeForkCluster}}. FORK clusters are available only on Unix-like operating systems, such as GNU/Linux and macOS. They are not supported on Microsoft Windows; on Windows, \code{parallel="parallel"} is automatically changed to \code{parallel="parallelWin"}. \cr

-)\kbd{parallelWin}: parallel computations using a \sQuote{PSOCK} cluster created with \code{\link[parallel]{makeCluster}}. PSOCK clusters are the only parallel cluster option supported by \pkg{parallel} on Microsoft Windows, and they also work on Unix-like systems. \cr

For both \code{parallel="parallel"} and \code{parallel="parallelWin"}, \code{fn.name="hydromod"} and \code{fn.name="hydromodInR"} model runs are distributed one parameter set at a time using the selected cluster. Ordinary R objective functions are evaluated row-wise with the \code{\link[parallel]{parRapply}} function of the \pkg{parallel} package. \cr

On Unix-like systems, FORK clusters are usually faster than PSOCK clusters because workers inherit the parent R session and require less explicit object/package export. PSOCK clusters are more portable but usually have higher startup and data-transfer overhead, especially when large objects must be sent to workers.
}
  \item{par.nnodes}{
OPTIONAL. Used only when \code{parallel!='none'} \cr

numeric, indicates the number of cores/CPUs to be used in the local multi-core machine, or the number of nodes to be used in the network cluster. By default \code{par.nnodes} is set to the amount of cores detected by \code{\link[parallel]{detectCores}}.
}
  \item{par.pkgs}{
OPTIONAL. Used only when \code{parallel='parallelWin'} \cr

list of package names (as characters) that need to be loaded on each PSOCK worker for allowing the objective function \code{fn} to be evaluated.
}
  \item{par.env}{
OPTIONAL. Used only when \code{parallel!='none'} \cr

environment from which the objects required by the objective function \code{fn}, \code{model.FUN}, \code{out.FUN}, or \code{gof.FUN} are exported to PSOCK workers. By default, the environment from which \code{verification} is called is used.
}
  \item{par.export}{
OPTIONAL. Used only when \code{parallel!='none'} \cr

character vector with the names of additional objects to be exported from \code{par.env} to PSOCK workers. By default, all functions in \code{par.env} are exported.
}
  }
}

\value{
A list of four elements: \cr 

\item{gofs}{numeric with the goodness-of-fit values corresponding to each one of the parameter sets provided in \code{par}}.

\item{model.values}{data.frame with the model outputs corresponding to each one of the parameter sets provided in \code{par}}.

\item{best.param}{numeric with the parameter values of the "best" paraemter set found during the verification period}.

\item{best.gof}{numeric with the goodness-of-fit of the "best" parameter set found during the verification period}

}

\author{
Mauricio Zambrano-Bigiarini, \email{mzb.devel@gmail.com}
}

\seealso{
\code{\link{hydromod}}, \code{\link{hydromodInR}}
}

\examples{
## R-external model verification with parameter metadata in 'drty.in'.
## The shell script used here is skipped on Windows.
if (.Platform$OS.type != "windows") {

  model.drty <- tempfile("verification-model-")
  drty.in    <- file.path(model.drty, "PSO.in")
  drty.out   <- tempfile("verification-out-")
  dir.create(drty.in, recursive=TRUE)

  writeLines("      0.00", file.path(model.drty, "input.txt"))

  writeLines(
    c("ParameterNmbr ParameterName MinValue MaxValue",
      "1 P1 0 10"),
    file.path(drty.in, "ParamRanges.txt")
  )

  writeLines(
    c("ParameterNmbr ParameterName Filename Row.Number Col.Start Col.End Decimals",
      "1 P1 input.txt 1 1 10 2"),
    file.path(drty.in, "ParamFiles.txt")
  )

  exe.fname <- file.path(model.drty, "run.model")
  writeLines(c("#!/bin/sh", "exit 0"), exe.fname)
  Sys.chmod(exe.fname, mode="0755")

  read_model_output <- function(file) {
    value <- as.numeric(trimws(readLines(file, warn=FALSE)[1]))
    c(value, value + 1)
  }

  sum_gof <- function(sim, obs) {
    sum(sim - obs)
  }

  out <- verification(
    fn="hydromod",
    par=matrix(c(1.25, 2.50), ncol=1),
    control=list(drty.in=drty.in, drty.out=drty.out, verbose=FALSE),
    model.FUN=hydromod,
    model.FUN.args=list(
      param.files="ParamFiles.txt",
      param.ranges="ParamRanges.txt",
      model.drty=model.drty,
      exe.fname=exe.fname,
      out.FUN=read_model_output,
      out.FUN.args=list(file=file.path(model.drty, "input.txt")),
      gof.FUN=sum_gof,
      obs=c(0, 0),
      verbose=FALSE
    )
  )

  out$gofs
}
}

\keyword{optimize}
\keyword{files}% __ONLY ONE__ keyword per line
